Fixed backup issue (due to sign extension).
git-svn-id: http://svn.tilp.info/repos/tilp@1570

+++ libticalcs/branches/ticalcs-2/src/calc_83.c
@@ -221,9 +221,9 @@ static int		recv_backup	(CalcHandle* handle, BackupContent* content)
 	TRYF(ti82_recv_ACK(&unused));
 
 	TRYF(ti82_recv_VAR(&(content->data_length1), &content->type, varname));
-	content->data_length2 = varname[0] | (varname[1] << 8);
-	content->data_length3 = varname[2] | (varname[3] << 8);
-	content->mem_address = varname[4] | (varname[5] << 8);
+	content->data_length2 = (uint8_t)varname[0] | ((uint8_t)varname[1] << 8);
+	content->data_length3 = (uint8_t)varname[2] | ((uint8_t)varname[3] << 8);
+	content->mem_address  = (uint8_t)varname[4] | ((uint8_t)varname[5] << 8);
 	TRYF(ti82_send_ACK());
 
 	TRYF(ti82_send_CTS());
